second fix for qr code login url creation
The mobile app does some crazy things with the URL. I think most of it
just boils down to picking the right space. It does it automatically by
matching the URL from the QR code with a list of spaces that it fetches
from PRONOTE. Otherwise it prompts the user to choose.

We're just prepending "mobile." if the URL doesn't have the prefix.

closes #299 (for real this time)