add 2FA support
Implement support for accounts with higher security enabled (from now on
2FA, even though it is not really 2FA in all cases).

For 2FA-enabled accounts, pronotepy needs to register itself as a device
on the account, which may require the account PIN code. The account PIN
and device name can be passed to the client constructor:

  import pronotepy
  client = pronotepy.Client(..., account_pin="1111", device_name="ppy")

When the device is registered, it is uniquely identified using the
client identifier. It is available on the client instance as
Client.client_identifier.

To make logging in easier, credentials can be exported using a new
method: Client.export_credentials. It will generate a dict with all the
necessary details. It can be directly passed in for as login parameters:

  creds = client.export_credentials()
  client = pronotepy.Client.token_login(**creds)

When the client tries to log in, but can't because of 2FA, MFAError is
raised.